[D32530.96705] [SVE] Scalable Vector IR Type
Differential Revision: https://reviews.llvm.org/D32530?id=96705

+++ llvm/lib/IR/LLVMContextImpl.h
@@ -1189,7 +1189,7 @@ public:
   unsigned NamedStructTypesUniqueID;
     
   DenseMap<std::pair<Type *, uint64_t>, ArrayType*> ArrayTypes;
-  DenseMap<std::pair<Type *, unsigned>, VectorType*> VectorTypes;
+  DenseMap<std::pair<Type *, uint64_t>, VectorType*> VectorTypes;
   DenseMap<Type*, PointerType*> PointerTypes;  // Pointers in AddrSpace = 0
   DenseMap<std::pair<Type*, unsigned>, PointerType*> ASPointerTypes;
